+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Also I copied this intro verbatim from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Changed:
+* Sublime Text is now disabled by default, since ST4 is able to sync with system dark-mode
 
 ## [v0.13.1 - 2021-11-05]
 ### Fixed: